requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23
protobuf==3.14.0
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
